requirements.txt
flask==2.3.3
flask-socketio==5.3.6
flask-cors==4.0.0
opencv-python-headless==4.8.0.74
numpy==1.24.3
scikit-learn==1.3.0
pillow==10.0.0
matplotlib==3.7.1
imutils==0.5.4